Split dev build workflow to work nicer with forks
Cron trigger workflows are disabled on forks by default so this change
adds a reusable workflow that is called by two workflows:

- one that works as before to build dev on push / PR
- one that runs the scheduled build

It also now builds airplay1 and airplay2 in parallel and changes the
name of the built image to match the repo name for forks.

Overall this should result in no actual change on the original repo
but should work more nicely with forks.